changed type of i from int to size_t

+++ myname.c
@@ -2,7 +2,7 @@
 #include <string.h>
 
 static char *program_path;
-static int i;
+static size_t i;
 
 int main(int argc, char **argv)
 {
@@ -12,7 +12,7 @@ int main(int argc, char **argv)
     program_path += 2;
   else if (program_path[0] == '/')
     {
-      for (i = (int)strlen(program_path) - 1; i >= 0; --i)
+      for (i = strlen(program_path) - 1; i >= 0; --i)
         if (program_path[i] == '/')
             break;
       program_path += ++i;
